Add error log when credentials cannot be identified (#72)
Add a private member function to the Configuration class
to be called after gathering any available local configuration
values. The function checks to see if at least one of the three credential
fields has been set. If this is not the case then the function will write
an error log telling the user that credentials have not been
supplied and authorization with the Stackdriver API will not be attainable
until such time as they are provided.