The service defaults to the function name on GCF (#70)
If the user has specified a service name, then that value will be 
used as the service name.  Otherwise, if the code is running on 
GCF, the name of the function being run will be used as the 
service name.  Last, if neither of those two conditions are true, 
the GCP service name will be used.

PR-URL: https://github.com/GoogleCloudPlatform/cloud-errors-nodejs/pull/70