Add error message builder and parse capability (#74)
Add a function exposed on the top-level export
of the library which allows for creation of new
ErrorMessage class instances for direct use and
modification by application code.

Additionally, add capability to the manual handler
to specifically accept error message class instances
and not apply general object handling. This means
that if a user supplies only an error message instance
the instance will be accepted wholesale and queued for
submission to the service without modification.